requirements.txt
py-discord-html-transcripts==2.1.1
paypalrestsdk==1.13.3
aiosqlite==0.21.0
discord.py==2.4.0
htmlmin==0.1.12
pyyaml==6.0.2
qrcode==8.0